tests: updated test code for objects rule
FossilOrigin-Name: 0a1cd960e6ef6153b907490c629dcd895722146bee5c06c5564396589969c365

+++ tests/valid/objects.nim
@@ -1,6 +1,6 @@
 # Count: 0
-# Negative count: 1
+# Negative count: 2
 
 type test = object
   a: char
-  b, c*: int8
+  b, c*: int
